+++ kern/gui.err
@@ -1,2 +1,2 @@
-gui/gui.c(405): Warning! W308: The function 'c_keyboard_loop' without prototyped parameters called
-gui/gui.c(405): Note! I2002: 'c_keyboard_loop' defined in: gui/../io.h(19)
+gui/gui.c(332): Warning! W308: The function 'c_keyboard_loop' without prototyped parameters called
+gui/gui.c(332): Note! I2002: 'c_keyboard_loop' defined in: gui/../io.h(19)
